Fix network API servers TLS
network API servers TLS was broken, because of an error while copying
all the manifests and Makefile from the main KOS repo. The error
resulted in the network API servers using always the same CA bundle
of certificates as opposed to using the one that gets generated on
every `make clean deploy`.